Sometimes, you need to handle patches more than one database.
In fact, with e-commerce sites, this is quite normal.

Of course, changes to those databases are frequently dependent on each other.

Wouldn't it be cool if AutoPatch was able to orchestrate patch application across many distributed datasources?

Well, it nearly can. The single-system case has been broken up a little bit to expose some of its internals so it can be controlled by an orchestrating algorithm, and configuration for this idea is implemented and tested now.

Global ordering or all patches is next, then patch issuance