o  Build system support for an "optional" package including application server bindings o  Changed the way junit dependency is handled to use environmental one, but check first o  Bumping build number to reflect improved support for dependency injecting containers
This last change may or may not break the Hongkong build.

I'm actually afraid it will, but I'm not sure how to protect against it. I'm going to just see if it breaks and fix it if so.

Most likely it will need the Dynamo DAS/lib/das.jar, though there may be other issues...