Updates dbunit dependency
The dbunit dependency coordinate has changed from a groupId of
`dbunit` to `org.dbunit`.  In addition, the minimal version in
the new group is 2.2.  This change encompasses both updates.

I made this change, because other projects that use newer versions
of dbunit can get conflicts if they depend on `org.dbunit` and the
autopatcher brings in a version of `dbunit`.